Delete images from Firestore

+++ app/src/main/java/com/steeshock/streetworkout/data/repository/implementation/firebase/FirebasePlacesRepository.kt
@@ -5,6 +5,8 @@ import androidx.lifecycle.LiveData
 import com.google.firebase.database.ktx.database
 import com.google.firebase.database.ktx.getValue
 import com.google.firebase.ktx.Firebase
+import com.google.firebase.storage.StorageException
+import com.google.firebase.storage.StorageReference
 import com.google.firebase.storage.ktx.storage
 import com.steeshock.streetworkout.common.Constants.FIREBASE_PATH
 import com.steeshock.streetworkout.data.database.PlacesDao
@@ -12,8 +14,10 @@ import com.steeshock.streetworkout.data.model.Place
 import com.steeshock.streetworkout.data.repository.interfaces.IPlacesRepository
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.tasks.await
+import java.lang.Exception
 import java.util.*
 import javax.inject.Inject
 import kotlin.coroutines.resume
@@ -37,7 +41,8 @@ open class FirebasePlacesRepository @Inject constructor(
             database.getReference("places").get().addOnSuccessListener {
                 for (child in it.children) {
                     val place = child.getValue<Place>()
-                    val isFavorite = allPlaces.value?.find { p -> p.placeId == place?.placeId }?.isFavorite
+                    val isFavorite =
+                        allPlaces.value?.find { p -> p.placeId == place?.placeId }?.isFavorite
                     place?.isFavorite = isFavorite == true
                     place?.let { p -> places.add(p) }
                 }
@@ -89,21 +94,21 @@ open class FirebasePlacesRepository @Inject constructor(
         placesDao.updatePlace(place)
     }
 
+    //TODO Do requests in parallel
     override suspend fun deletePlace(place: Place): Boolean {
         return suspendCoroutine { continuation ->
-            val database = Firebase.database(FIREBASE_PATH)
-            database.getReference("places").child(place.placeId).get()
-                .addOnSuccessListener { dataSnapshot ->
-                    dataSnapshot.ref.removeValue().addOnSuccessListener {
-                        CoroutineScope(Dispatchers.IO).launch {
-                            placesDao.deletePlace(place)
-                            continuation.resume(true)
-                        }
+            CoroutineScope(Dispatchers.IO).launch {
+                try {
+                    if (removePlaceRemote(place) && removeRelatedImagesRemote(place)) {
+                        placesDao.deletePlace(place)
+                        continuation.resume(true)
+                    } else {
+                        continuation.resume(false)
                     }
+                } catch (e: Exception) {
+                    continuation.resumeWithException(e)
                 }
-                .addOnFailureListener {
-                    continuation.resumeWithException(it)
-                }
+            }
         }
     }
 
@@ -122,6 +127,58 @@ open class FirebasePlacesRepository @Inject constructor(
         placesDao.getAllPlaces().apply {
             forEach { it.isFavorite = false }
             placesDao.insertAllPlaces(this)
+        }
+    }
+
+    private suspend fun removePlaceRemote(place: Place): Boolean {
+        return suspendCoroutine { continuation ->
+            val database = Firebase.database(FIREBASE_PATH)
+            database.getReference("places").child(place.placeId).get()
+                .addOnSuccessListener { dataSnapshot ->
+                    dataSnapshot.ref.removeValue().addOnSuccessListener {
+                        CoroutineScope(Dispatchers.IO).launch {
+                            continuation.resume(true)
+                        }
+                    }
+                }
+                .addOnFailureListener {
+                    continuation.resumeWithException(it)
+                }
+        }
+    }
+
+    /**
+     * Try to delete place's related images from Firebase Storage
+     * If images doesn't exist - Firebase returns error code -13010
+     */
+    private suspend fun removeRelatedImagesRemote(place: Place): Boolean {
+        return suspendCoroutine { continuation ->
+            Firebase.storage.reference.child(place.placeId).listAll()
+                .addOnSuccessListener { images ->
+                    CoroutineScope(Dispatchers.IO).launch {
+                        images.items.forEach { deleteSingleImage(it) }
+                        continuation.resume(true)
+                    }
+                }
+                .addOnFailureListener {
+                    if ((it as? StorageException)?.errorCode == StorageException.ERROR_OBJECT_NOT_FOUND) {
+                        continuation.resume(true)
+                    } else {
+                        continuation.resumeWithException(it)
+                    }
+                }
+        }
+    }
+
+    private suspend fun deleteSingleImage(storageReference: StorageReference): Boolean {
+        return suspendCoroutine { continuation ->
+            storageReference.delete()
+                .addOnSuccessListener {
+                    continuation.resume(true)
+                }
+                .addOnFailureListener {
+                    continuation.resumeWithException(it)
+                }
         }
     }
 }
